separate lattice variable config logib to new file
